data_analysis/requirements.txt
pandas==1.4.3
requests==2.28.1
python-dotenv==1.0.0
streamlit
snowflake-connector-python
pyarrow==8.0.0